refactor: move static content to separate files
