+++ content/specialties/ember.js
@@ -1,0 +1,33 @@
+export const head = {
+  title: 'Ember.js Enterprise App Development',
+  description:
+    "Learn how Ship Shape's top-tier enterprise app development experts can help your team build fast, reliable Ember.js applications.",
+  url: 'https://shipshape.io/services/ember-consulting/'
+};
+
+export const specialtyDetails = {
+  text: {
+    h2: 'All about Ember.js',
+    h3: 'Ember is great for large companies because of its:',
+    boxedText:
+      'Ember.js is an open-source JavaScript web framework. We love its strong opinions, which help engineers work faster and keep their work organized. ',
+    unboxedText: ''
+  },
+  benefits: [
+    { imgSrc: '/svgs/icons/modular.svg', description: 'Clear Organization' },
+    {
+      imgSrc: '/svgs/icons/chart.svg',
+      description: 'Scalable UI architecture'
+    },
+    { imgSrc: '/svgs/icons/retainer.svg', description: 'Simple updates' },
+    { imgSrc: '/svgs/icons/sail.svg', description: 'Commitment to stability' }
+  ]
+};
+
+export const testimonial = {
+  name: 'Roger Studner',
+  title: 'Chief Architect at Expel',
+  imgSrc: 'roger',
+  svgSrc: '/svgs/clients/expel-white.svg',
+  quote: `Ship Shape was a great aid to us because they were the voice of Ember best practices to keep our code base current, and they were able to deliver on critical features on our roadmap that we did not have the bandwidth to complete on time. I would absolutely recommend Ship Shape to any other business that is looking for an incredibly skilled partner to work closely with their teams to produce and deliver the highest value.`
+};
