+++ content/specialties/next.js
@@ -1,0 +1,51 @@
+export const head = {
+  title: 'Next.js Software Consultants & Developers',
+  description:
+    'As a proven leader in Next.js development, Ship Shape turns ideas into industry-leading products. ' +
+    'Contact us to get started.',
+  url: 'https://shipshape.io/services/next-consulting/'
+};
+
+export const specialtyDetails = {
+  text: {
+    h2: 'All about Next.js',
+    h3: 'Some of our favorite Next.js features include:',
+    boxedText:
+      'Teams who have worked in generic React apps can easily make the leap into Next.js. If you’re wondering “What is Next.js, and can my team transition?” Ship Shape’s experienced developers not only consult and develop in Next.js, but also train and mentor teams to maintain and improve their code over time.',
+    unboxedText: `Whether you’re dipping your toes into the app development waters and wondering, “What is Next.js?” or you already know the ropes and are exploring this open-source React framework, you’ve come to the right place. \n \n At Ship Shape, our veteran engineers recommend Next.js because it’s one of the most performant frameworks. With server-side rendering (SSR) and routing, TypeScript support, smart bundling, route pre-fetching, and more, it’s a great choice for production-ready web apps.'`
+  },
+  benefits: [
+    {
+      imgSrc: '/svgs/icons/image.svg',
+      description: 'Out-of-the-box internationalization features'
+    },
+    {
+      imgSrc: '/svgs/icons/mentor.svg',
+      description:
+        'Automatic image resizing and one-off, on-the-fly image generation'
+    },
+    {
+      imgSrc: '/svgs/icons/chart.svg',
+      description: 'Amazingly high performance'
+    },
+    {
+      imgSrc: '/svgs/icons/pwa.svg',
+      description:
+        'Zero config! Developing features is a simple “yarn dev” away'
+    }
+  ]
+};
+
+export const testimonial = {
+  name: 'Harley Sugarman',
+  title: 'Founder, CEO at Enigma',
+  imgSrc: 'harley',
+  svgSrc: '/svgs/clients/enigma-white.svg',
+  quote: `I loved working with Ship Shape.
+          They helped our small team build a product from the ground up
+          and were strong partners every step of the way. Their domain expertise
+          in Next.js and React helped lay the foundation for a scalable codebase,
+          and they were great at ensuring we adhered to engineering best practices.
+          Every member of the team was a pleasure to work with and (best of all)
+          we ended up launching our MVP on schedule!`
+};
