+++ content/specialties/serverless.js
@@ -1,0 +1,41 @@
+export const head = {
+  title: 'Serverless Software Consultants & Developers',
+  description:
+    'As a proven leader in serverless development, Ship Shape turns ideas into industry-leading products. ' +
+    'Contact us to get started.',
+  url: 'https://shipshape.io/services/serverless-consulting/'
+};
+
+export const specialtyDetails = {
+  text: {
+    h2: 'All about Serverless',
+    h3: 'This means your company can:',
+    boxedText:
+      'What is serverless? It’s a model that offloads server management to a third party, only charges you for actual usage, and makes it easy to auto-scale as demand increases. \n \n Why pay for computing power you aren’t using? With serverless frameworks, high server costs are a thing of the past.',
+    unboxedText:
+      'Smaller APIs, graph QL servers, front-end applications, microservices, and similar types of software work well within serverless frameworks. \n \n Working on something like this? Ship Shape’s first-rate software engineers know all the ins and outs. Whether you’re wondering, “What is serverless, and can it help my team?” or you’ve gotten started but need help taking it to the next level, we’ve got you covered.'
+  },
+  benefits: [
+    { imgSrc: '/svgs/icons/modular.svg', description: 'Clear Organization' },
+    {
+      imgSrc: '/svgs/icons/chart.svg',
+      description: 'Scalable UI architecture'
+    },
+    { imgSrc: '/svgs/icons/retainer.svg', description: 'Simple updates' },
+    { imgSrc: '/svgs/icons/sail.svg', description: 'Commitment to stability' }
+  ]
+};
+
+export const testimonial = {
+  name: 'Harley Sugarman',
+  title: 'Founder, CEO at Enigma',
+  imgSrc: 'harley',
+  svgSrc: '/svgs/clients/enigma.svg',
+  quote: `I loved working with Ship Shape. 
+          They helped our small team build a product from the ground up 
+          and were strong partners every step of the way. Their domain expertise 
+          in Next.js and React helped lay the foundation for a scalable codebase, 
+          and they were great at ensuring we adhered to engineering best practices. 
+          Every member of the team was a pleasure to work with and (best of all) 
+          we ended up launching our MVP on schedule!`
+};
