+++ content/specialties/tailwind.js
@@ -1,0 +1,51 @@
+export const head = {
+  title: 'Tailwind CSS Software Consultants & Developers',
+  description:
+    'As a proven leader in Tailwind CSS development, Ship Shape turns ideas into industry-leading products. ' +
+    'Contact us to get started.',
+  url: 'https://shipshape.io/services/tailwind-consulting/'
+};
+
+export const specialtyDetails = {
+  text: {
+    h2: 'All about Tailwind CSS',
+    h3: 'Ship Shape’s experienced engineers love Tailwind CSS because:',
+    boxedText:
+      'Much faster than with most CSS frameworks, Tailwind CSS provides almost all the tools you need to build a site without writing any custom styles.',
+    unboxedText: ''
+  },
+  benefits: [
+    {
+      imgSrc: '/svgs/icons/image.svg',
+      description: 'It matches mocks and designs with ease.'
+    },
+    {
+      imgSrc: '/svgs/icons/time.svg',
+      description:
+        'It automatically removes all unused CSS when building for production, which improves speed.'
+    },
+    {
+      imgSrc: '/svgs/icons/modular.svg',
+      description:
+        'Its usability is quite nimble compared to similar frameworks (like Bootstrap).'
+    },
+    {
+      imgSrc: '/svgs/icons/pwa.svg',
+      description: 'Prototyping is super fast with Tailwind UI components.'
+    }
+  ]
+};
+
+export const testimonial = {
+  name: 'Harley Sugarman',
+  title: 'Founder, CEO at Enigma',
+  imgSrc: 'harley',
+  svgSrc: '/svgs/clients/enigma.svg',
+  quote: `I loved working with Ship Shape.
+          They helped our small team build a product from the ground up
+          and were strong partners every step of the way. Their domain expertise
+          in Next.js and React helped lay the foundation for a scalable codebase,
+          and they were great at ensuring we adhered to engineering best practices.
+          Every member of the team was a pleasure to work with and (best of all)
+          we ended up launching our MVP on schedule!`
+};
