+++ pages/services/cloud-native-consulting.vue
@@ -103,49 +103,11 @@
 
 <script>
 import { generateMeta } from '~/utils/meta';
-
-const specialtyDetails = {
-  text: {
-    h2: 'All about Cloud Native',
-    h3: 'Ship Shape’s engineering staff prefers using a cloud-native approach because of its:',
-    boxedText:
-      'If you’re working on developing cloud-native applications or would like to explore the value they could bring to your team and customers, look no further. Our cloud-native specialists can help you build efficient and effective products today, lay a solid foundation for the future, and upskill your team so they can continue making positive progress long after our engagement ends.',
-    unboxedText:
-      'Cloud-native development is a methodology for building and running scalable applications in modern, dynamic environments. This platform-agnostic approach gives you flexibility so you don’t have to work with one vendor (such as AWS or Azure) and provides engineering teams with tools to set up a clear, effective software delivery pipeline so they can deliver battle-tested, secure solutions faster.'
-  },
-  benefits: [
-    {
-      imgSrc: '/svgs/icons/retainer.svg',
-      description: 'Fast, transparent software delivery cycle'
-    },
-    {
-      imgSrc: '/svgs/icons/speed.svg',
-      description: 'Ability to deliver value to customers even faster'
-    },
-    {
-      imgSrc: '/svgs/icons/secure.svg',
-      description: 'Tested, safe, and secure deployment'
-    },
-    {
-      imgSrc: '/svgs/icons/chart.svg',
-      description: 'Rapid scalability'
-    }
-  ]
-};
-
-const testimonial = {
-  name: 'Harley Sugarman',
-  title: 'Founder, CEO at Enigma',
-  imgSrc: 'harley',
-  svgSrc: '/svgs/clients/enigma.svg',
-  quote: `I loved working with Ship Shape.
-          They helped our small team build a product from the ground up
-          and were strong partners every step of the way. Their domain expertise
-          in Next.js and React helped lay the foundation for a scalable codebase,
-          and they were great at ensuring we adhered to engineering best practices.
-          Every member of the team was a pleasure to work with and (best of all)
-          we ended up launching our MVP on schedule!`
-};
+import {
+  head,
+  specialtyDetails,
+  testimonial
+} from '~/content/specialties/cloud-native.js';
 
 export default {
   speedkitComponents: {
@@ -155,15 +117,10 @@ export default {
     WaveMoreLinks: () => import('@/components/WaveMoreLinks')
   },
   data() {
-    return { testimonial, specialtyDetails };
+    return { head, testimonial, specialtyDetails };
   },
   head() {
-    const title = 'Next.js Software Consultants & Developers';
-    const description =
-      'As a proven leader in Next.js development, Ship Shape turns ideas into industry-leading products. ' +
-      'Contact us to get started.';
-    const url = 'https://shipshape.io/services/next-consulting/';
-
+    const { title, description, url } = head;
     return generateMeta(title, description, url);
   }
 };
