+++ pages/services/ember-consulting.vue
@@ -234,33 +234,11 @@
 
 <script>
 import { generateMeta } from '~/utils/meta';
-
-const specialtyDetails = {
-  text: {
-    h2: 'All about Ember.js',
-    h3: 'Ember is great for large companies because of its:',
-    boxedText:
-      'Ember.js is an open-source JavaScript web framework. We love its strong opinions, which help engineers work faster and keep their work organized. ',
-    unboxedText: ''
-  },
-  benefits: [
-    { imgSrc: '/svgs/icons/modular.svg', description: 'Clear Organization' },
-    {
-      imgSrc: '/svgs/icons/chart.svg',
-      description: 'Scalable UI architecture'
-    },
-    { imgSrc: '/svgs/icons/retainer.svg', description: 'Simple updates' },
-    { imgSrc: '/svgs/icons/sail.svg', description: 'Commitment to stability' }
-  ]
-};
-
-const testimonial = {
-  name: 'Roger Studner',
-  title: 'Chief Architect at Expel',
-  imgSrc: 'roger',
-  svgSrc: '/svgs/clients/expel-white.svg',
-  quote: `Ship Shape was a great aid to us because they were the voice of Ember best practices to keep our code base current, and they were able to deliver on critical features on our roadmap that we did not have the bandwidth to complete on time. I would absolutely recommend Ship Shape to any other business that is looking for an incredibly skilled partner to work closely with their teams to produce and deliver the highest value.`
-};
+import {
+  head,
+  specialtyDetails,
+  testimonial
+} from '~/content/specialties/ember.js';
 
 export default {
   speedkitComponents: {
@@ -272,14 +250,11 @@ export default {
   },
 
   data() {
-    return { testimonial, specialtyDetails };
+    return { head, testimonial, specialtyDetails };
   },
 
   head() {
-    const title = 'Ember.js Enterprise App Development';
-    const description =
-      "Learn how Ship Shape's top-tier enterprise app development experts can help your team build fast, reliable Ember.js applications.";
-    const url = 'https://shipshape.io/services/ember-consulting/';
+    const { title, description, url } = head;
 
     return generateMeta(title, description, url);
   }
