+++ pages/services/nuxt-consulting.vue
@@ -214,43 +214,11 @@
 
 <script>
 import { generateMeta } from '~/utils/meta';
-const specialtyDetails = {
-  text: {
-    h2: 'All about Nuxt.js',
-    h3: 'Nuxt works well for apps big and small thanks to its:',
-    boxedText:
-      'Nuxt.js is an open-source framework used to build Vue.js web applications. It’s a performant system that makes creating universal apps easier because of its reduced routing complexity, giving you SEO benefits right out of the box.',
-    unboxedText: `At Ship Shape, we love Nuxt.js because it’s still an opinionated framework (though not quite as opinionated as Ember.js), and it has features that strip out unused code to make your app fast and lightweight. You’ll enjoy its approachable syntax, easy-to-read read templates, and more. \n \n Whether you're building a brand new web app or elevating an existing experience with the latest best practices, our first-class engineers can refine or create a Nuxt.js app while training your team in the process.`
-  },
-  benefits: [
-    {
-      imgSrc: '/svgs/icons/modular.svg',
-      description: 'Modular architecture, which speeds up development'
-    },
-    {
-      imgSrc: '/svgs/icons/native-apps.svg',
-      description:
-        'Server-side rendered (SSR) to deliver HTML, rather than JavaScript'
-    },
-    {
-      imgSrc: '/svgs/icons/static.svg',
-      description:
-        'Static Site Generation to net SEO benefits without needing a server'
-    }
-  ]
-};
-
-const testimonial = {
-  name: 'Donald Piret',
-  title: 'Director of Engineering at Sephora',
-  imgSrc: 'donald',
-  svgSrc: '/svgs/clients/sephora-white.svg',
-  quote: `Ship Shape delivered beyond our expectations. They pointed out things
-          we had never even thought of or knew about. Our team learned a huge
-          amount of new information and new tricks, and we still have them on
-          retainer for very specific questions. Plus, Ship Shape is the
-          friendliest team of engineers I've worked with.`
-};
+import {
+  head,
+  specialtyDetails,
+  testimonial
+} from '~/content/specialties/nuxt.js';
 
 export default {
   speedkitComponents: {
@@ -260,14 +228,10 @@ export default {
     WaveMoreLinks: () => import('@/components/WaveMoreLinks')
   },
   data() {
-    return { testimonial, specialtyDetails };
+    return { head, testimonial, specialtyDetails };
   },
   head() {
-    const title = 'Nuxt.js Software Consultants & Developers';
-    const description =
-      'As a proven leader in Nuxt development, Ship Shape turns ideas into industry-leading products. ' +
-      'Contact us to get started.';
-    const url = 'https://shipshape.io/services/nuxt-consulting/';
+    const { title, description, url } = head;
 
     return generateMeta(title, description, url);
   }
